Move Tika content handlers to Common
